Roughly implement flexbox design layout
Have added multiple flexbox containers to the layout. You can see the
barebone structure of the page and it roughly matches the target design.
The logo and navbar is correct, as well as the info section but, I'm
having some issues with the hero text and image box as there is spacing
around the hero header and paragraph and I'm not sure why. Adjusting
justify-content isn't working so I will have to go back and review.

Other than that, I have added some of the text colors and sizes as well
as centered the text on the footer.