Adjust margin and padding on the flex containers
Margin and padding adjustments were done to match the design. Also added
another flex container to the signup that enclosed the call to action
and text underneath it. I did this so that the signup button would
appear to the right of the call text. I also added self-align the author
of the quote using `flex-end`.