Narrow the main content of the page
I increased the padding and margin on a few elements in order to
centralize the content a bit more because I realized that the items were
far too spaced when I went to fullscreen. I don't know what device
width the target design was done on but seeing as I have a wider monitor
I elected to not work in full screen but still narrow it more as this
is not meant to be a mobile page design yet.

There was some styling added to the info header as well.